All cutscene implementation details work. Level1 has the cutscene inserted in the storyboard for reference.

+++ SecurityGame/BGKVCutsceneModel.h
@@ -12,7 +12,7 @@
 
 - (id) initWithLevel: (int) level;
 
-- (int) getMaxDialogueLevel: (int) level;
+- (NSUInteger) getMaxDialogueLevel: (int) level;
 
 - (NSString*)getStringForDialogue: (int) dialogue;
